+++ pkg/service/service.go
@@ -9,7 +9,9 @@ import (
 
 	devfile "github.com/devfile/api/v2/pkg/apis/workspaces/v1alpha2"
 	"github.com/devfile/library/pkg/devfile/parser/data/v2/common"
+	parsercommon "github.com/devfile/library/pkg/devfile/parser/data/v2/common"
 	"github.com/openshift/odo/pkg/kclient"
+	"github.com/openshift/odo/pkg/log"
 	"github.com/openshift/odo/pkg/odo/util/validation"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -308,6 +310,10 @@ func ListOperatorServices(client *kclient.Client) ([]unstructured.Unstructured, 
 
 	// First let's get the list of all the operators in the namespace
 	csvs, err := client.ListClusterServiceVersions()
+	if err == kclient.ErrNoSuchOperator {
+		return nil, nil, err
+	}
+
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "Unable to list operator backed services")
 	}
@@ -692,6 +698,43 @@ func IsCSVSupported() (bool, error) {
 	return client.GetKubeClient().IsCSVSupported()
 }
 
+// IsDefined checks if a service with the given name is defined in a DevFile
+func IsDefined(name string, devfileObj parser.DevfileObj) (bool, error) {
+	components, err := devfileObj.Data.GetComponents(common.DevfileOptions{})
+	if err != nil {
+		return false, err
+	}
+	for _, c := range components {
+		if c.Name == name {
+			return true, nil
+		}
+	}
+	return false, nil
+}
+
+// ListDevfileServices returns the names of the services defined in a Devfile
+func ListDevfileServices(devfileObj parser.DevfileObj) ([]string, error) {
+	if devfileObj.Data == nil {
+		return nil, nil
+	}
+	components, err := devfileObj.Data.GetComponents(common.DevfileOptions{
+		ComponentOptions: parsercommon.ComponentOptions{ComponentType: devfile.KubernetesComponentType},
+	})
+	if err != nil {
+		return nil, err
+	}
+	var services []string
+	for _, c := range components {
+		var u unstructured.Unstructured
+		err = yaml.Unmarshal([]byte(c.Kubernetes.Inlined), &u)
+		if err != nil {
+			return nil, err
+		}
+		services = append(services, strings.Join([]string{u.GetKind(), c.Name}, "/"))
+	}
+	return services, nil
+}
+
 // AddKubernetesComponentToDevfile adds service definition to devfile as an inlined Kubernetes component
 func AddKubernetesComponentToDevfile(crd, name string, devfileObj parser.DevfileObj) error {
 	err := devfileObj.Data.AddComponents([]devfile.Component{{
@@ -814,15 +857,29 @@ func (d *DynamicCRD) AddComponentLabelsToCRD(labels map[string]string) {
 	metaMap["labels"] = labels
 }
 
-// CreateServiceFromKubernetesInlineComponents creates service(s) from Kubernetes Inlined component in a devfile
-func CreateServiceFromKubernetesInlineComponents(client *kclient.Client, k8sComponents []devfile.Component, labels map[string]string) ([]string, error) {
-	if len(k8sComponents) == 0 {
-		// if there's no Kubernetes Inlined component, there's nothing to do.
-		return []string{}, nil
+// PushServiceFromKubernetesInlineComponents updates service(s) from Kubernetes Inlined component in a devfile by creating new ones or removing old ones
+func PushServiceFromKubernetesInlineComponents(client *kclient.Client, k8sComponents []devfile.Component, labels map[string]string) error {
+
+	created := []string{}
+	deleted := []string{}
+
+	deployed := map[string]struct{}{}
+
+	deployedServices, _, err := ListOperatorServices(client)
+	if err != nil && err != kclient.ErrNoSuchOperator {
+		// We ignore ErrNoSuchOperator error as we can deduce Operator Services are not installed
+		return err
+	}
+	for _, svc := range deployedServices {
+		name := svc.GetName()
+		kind := svc.GetKind()
+		deployedLabels := svc.GetLabels()
+		if deployedLabels[applabels.ManagedBy] == "odo" && deployedLabels[componentlabels.ComponentLabel] == labels[componentlabels.ComponentLabel] {
+			deployed[kind+"/"+name] = struct{}{}
+		}
 	}
 
 	// create an object on the kubernetes cluster for all the Kubernetes Inlined components
-	var services []string
 	for _, c := range k8sComponents {
 		// get the string representation of the YAML definition of a CRD
 		strCRD := c.Kubernetes.Inlined
@@ -831,12 +888,12 @@ func CreateServiceFromKubernetesInlineComponents(client *kclient.Client, k8sComp
 		d := NewDynamicCRD()
 		err := yaml.Unmarshal([]byte(strCRD), &d.OriginalCRD)
 		if err != nil {
-			return []string{}, err
+			return err
 		}
 
 		cr, csv, err := GetCSV(client, d.OriginalCRD)
 		if err != nil {
-			return []string{}, err
+			return err
 		}
 
 		var group, version, kind, resource string
@@ -844,7 +901,7 @@ func CreateServiceFromKubernetesInlineComponents(client *kclient.Client, k8sComp
 			if crd.Kind == cr {
 				group, version, kind, resource, err = getGVKRFromCR(crd)
 				if err != nil {
-					return []string{}, err
+					return err
 				}
 				break
 			}
@@ -853,22 +910,61 @@ func CreateServiceFromKubernetesInlineComponents(client *kclient.Client, k8sComp
 		// add labels to the CRD before creation
 		d.AddComponentLabelsToCRD(labels)
 
+		crdName, ok := getCRDName(d.OriginalCRD)
+		if !ok {
+			continue
+		}
+
+		delete(deployed, cr+"/"+crdName)
+
 		// create the service on cluster
 		err = client.CreateDynamicResource(d.OriginalCRD, group, version, resource)
 		if err != nil {
 			if strings.Contains(err.Error(), "already exists") {
 				// this could be the case when "odo push" was executed after making change to code but there was no change to the service itself
 				// TODO: better way to handle this might be introduced by https://github.com/openshift/odo/issues/4553
-				err = nil
 				continue // this ensures that services slice is not updated
 			} else {
-				return []string{}, err
+				return err
 			}
 		}
 
 		name, _ := d.GetServiceNameFromCRD() // ignoring error because invalid yaml won't be inserted into devfile through odo
-		services = append(services, strings.Join([]string{kind, name}, "/"))
+		created = append(created, strings.Join([]string{kind, name}, "/"))
 	}
 
-	return services, nil
+	for key := range deployed {
+		err = DeleteOperatorService(client, key)
+		if err != nil {
+			return err
+
+		}
+		deleted = append(deleted, key)
+	}
+
+	if len(created) == 1 {
+		log.Successf("Created service %q on the cluster; refer %q to know how to link it to the component", created[0], "odo link -h")
+	} else if len(created) > 1 {
+		log.Successf("Created services %q on the cluster; refer %q to know how to link them to the component", strings.Join(created, ", "), "odo link -h")
+	}
+
+	if len(deleted) == 1 {
+		log.Successf("Deleted service %q from the cluster", deleted[0])
+	} else if len(deleted) > 1 {
+		log.Successf("Deleted services %q from the cluster", strings.Join(deleted, ", "))
+	}
+
+	return nil
+}
+
+func getCRDName(crd map[string]interface{}) (string, bool) {
+	metadata, ok := crd["metadata"].(map[string]interface{})
+	if !ok {
+		return "", false
+	}
+	name, ok := metadata["name"].(string)
+	if !ok {
+		return "", false
+	}
+	return name, true
 }