+++ Changelog.md
@@ -1,5 +1,19 @@
 #  Odo Changelog
 
+## 2.x
+
+### Feature/Enhancements
+
+- `odo delete` removes Operator backed services from the Devfile only, without undeploying them from the cluster ([#4761](https://github.com/openshift/odo/pull/4761))
+- `odo service list` lists Operator backed services in Devfile and/or deployed into the cluster and their state (Pushed, Not pushed, Deleted locally) ([#4761](https://github.com/openshift/odo/pull/4761))
+- `odo push` undeploys Operator backed services marked as managed by the current devfile not present in this devfile anymore ([#4761](https://github.com/openshift/odo/pull/4761))
+
+### Bug Fixes
+
+### Tests
+
+### Documentation
+
 ## 2.2.2
 
 ### Feature/Enhancements